==> src/2. Data Processing - Labeling - Splitting/3. data_explore_labeling.ipynb <==
## Combining Organization and Replies from batch

import pandas as pd

path = "../../data/3.1. Annotated"
save_path = "../../data/3.2. Annotated Combined"
organization_file = [
    "organization_1.xlsx",
    "organization_2.xlsx",
    "organization_3.xlsx",
    "organization_4.xlsx",
    "organization_5.xlsx",
]
replies_file = [
    "replies_100_1.xlsx",
    "replies_100_2.xlsx",
    "replies_100_3.xlsx",
    "replies_100_4.xlsx",
    "replies_100_5.xlsx",
]

organization_df = pd.concat(
    [
        pd.read_excel(f"{path}/{filename}").replace(
            {"\n": " ", "\r": "", "_x000D_": " "}, regex=True
        )
        for filename in organization_file
    ]
)
replies_df = pd.concat(
    [
        pd.read_excel(f"{path}/{filename}").replace(
            {"\n": " ", "\r": "", "_x000D_": " "}, regex=True
        )
        for filename in replies_file
    ]
)

organization_df.to_csv(f"{save_path}/organization.csv", index=False)
replies_df.to_csv(f"{save_path}/replies.csv", index=False)


## Check assumption
How many non-headline from organization files?  
How many headline from replies files?

num_of_organization = len(organization_df)
num_of_replies = len(replies_df)

num_of_headline_organization = len(organization_df[organization_df['labels (Non-Headline 0 / Headline 1)'] == 1])
num_of_non_headline_replies = len(replies_df[replies_df['labels (Non-Headline 0 / Headline 1)'] == 0])

print("With an assumption that organization are all headline and replies are all non-headline")

print("num_of_organization (assumption all headline):", num_of_organization)
print("num_of_replies (assumption all non-headline):", num_of_replies)

print("num_of_organization (actual headline):", num_of_headline_organization)
print("num_of_replies (actual non-headline):", num_of_non_headline_replies)

print("num_of_organization (wrong assumed headline):", num_of_organization - num_of_headline_organization)
print("num_of_replies (wrong assumed non-headline):", num_of_replies - num_of_non_headline_replies)
