==> src/2. Data Processing - Labeling - Splitting/5. data_explore_preprocesing.ipynb <==
import pandas as pd

scenario = [i for i in range(1, 7)]
data_path = "../../data/4. Processed"

data = {}
for i in scenario:
    full_path = f"{data_path}/{i}"
    train_test_data = {}
    train_test_data['train'] = pd.read_csv(f"{data_path}/{i}/train.csv")
    train_test_data['test'] = pd.read_csv(f"{data_path}/{i}/test.csv")
    data[i] = train_test_data

data[1]['train']
